+++ components/Tag/CHANGELOG.md
@@ -1,5 +1,17 @@
 # @shallot-ui/tag
 
+## 0.4.1
+
+### Patch Changes
+
+- adds full height and width style props
+- Updated dependencies
+  - @shallot-ui/text@0.4.1
+  - @shallot-ui/box@0.4.1
+  - @shallot-ui/row@0.4.1
+  - @shallot-ui/theme@0.27.1
+  - @shallot-ui/core@5.0.1
+
 ## 0.4.0
 
 ### Minor Changes